refactor(web): 구조 변경
