+++ apps/web/src/app/product/[id]/reservation-form/_form/ReservationFormContent.tsx
@@ -2,7 +2,7 @@
 
 import { type FormEvent } from 'react';
 import { ClientFooter, SchedulePickerDrawers } from '../components';
-import { useReservationCopyForm } from '../hooks';
+import useReservationCopyForm from '../hooks/useReservationCopyForm';
 import RESERVATION_FORM_MOCK from '../mock/reservationForm.mock';
 import AdditionalRequestSection from './AdditionalRequestSection';
 import ReservationDetailsSection from './ReservationDetailsSection';
