+++ apps/web/src/app/product/[id]/reservation-form/constants/copyForm.ts
