--- apps/web/src/app/product/[id]/reservation-form/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useReservationCopyForm } from './useReservationCopyForm';
-export type { ReservationCopyFormModel } from './useReservationCopyForm';
